Added implementation for MA method access

+++ Ohouse/src/plugins/ofed2/omavtwodelegate.py
@@ -42,10 +42,12 @@ class OMAv2Delegate(GMAv2DelegateBase):
         the resource manager.
         """
         if (type_=='KEY'):
+            self._delegate_tools.check_if_authorized(credentials, certificate, 'CREATE', 'KEY')
             self._delegate_tools.object_creation_check(fields, self._key_whitelist)
             self._delegate_tools.object_consistency_check(type_, fields)
             return self._member_authority_resource_manager.create_key(certificate, credentials, fields, options)
         elif (type_ =='MEMBER'):
+            self._delegate_tools.check_if_authorized(credentials, certificate, 'CREATE', 'SYSTEM_MEMBER')
             return self._member_authority_resource_manager.register_member(certificate, credentials, fields, options)
         else:
             raise gfed_ex.GFedv2NotImplementedError("__No create method found for object type: " + str(type_))
@@ -57,10 +59,12 @@ class OMAv2Delegate(GMAv2DelegateBase):
         the resource manager.
         """
         if (type_=='MEMBER'):
+            self._delegate_tools.check_if_ma_info_update_authorized(credentials, certificate, 'SYSTEM_MEMBER', urn)
             self._delegate_tools.object_update_check(fields, self._member_whitelist)
             self._delegate_tools.object_consistency_check(type_, fields)
             return self._member_authority_resource_manager.update_member(urn, certificate, credentials, fields, options)
         elif (type_=='KEY'):
+            self._delegate_tools.check_if_ma_info_update_authorized(credentials, certificate, type_, urn)
             self._delegate_tools.object_update_check(fields, self._key_whitelist)
             self._delegate_tools.object_consistency_check(type_, fields)
             return self._member_authority_resource_manager.update_key(urn, certificate, credentials, fields, options)
@@ -73,6 +77,7 @@ class OMAv2Delegate(GMAv2DelegateBase):
         using the resource manager.
         """
         if (type_=='KEY'):
+            self._delegate_tools.check_if_ma_info_update_authorized(credentials, certificate, type_, urn)
             return self._member_authority_resource_manager.delete_key(urn, certificate, credentials, options)
         else:
             raise gfed_ex.GFedv2NotImplementedError("No delete method found for object type: " + str(type_))
@@ -83,8 +88,10 @@ class OMAv2Delegate(GMAv2DelegateBase):
         using the resource manager.
         """
         if (type_=='MEMBER'):
+            self._delegate_tools.check_if_authorized(credentials, certificate, 'LOOKUP', 'SYSTEM_MEMBER')
             return self._delegate_tools.to_keyed_dict(self._member_authority_resource_manager.lookup_member(certificate, credentials, match, filter_, options), "MEMBER_URN")
         elif (type_=='KEY'):
+            self._delegate_tools.check_if_authorized(credentials, certificate, 'LOOKUP', 'KEY')
             return self._delegate_tools.to_keyed_dict(self._member_authority_resource_manager.lookup_key(certificate, credentials, match, filter_, options), "KEY_ID")
         else:
             raise gfed_ex.GFedv2NotImplementedError("No lookup method found for object type: " + str(type_))
